move converter Diagnostic into its own package
